requirements.txt
numpy=1.16.2
pandas=0.18.1
sqlalchemy=1.4.12
mysql-connector-python=8.0.17
PyYAML=3.12
TA-Lib=0.4.17